# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Indian Cinema Tracker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tmdb/search:
    get:
      operationId: searchTmdb
      tags: [tmdb]
      summary: Search TMDB for movies
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: region
          in: query
          schema:
            type: string
            default: "IN"
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TmdbMovie"
...
  /tmdb/upcoming:
    get:
      operationId: getUpcomingReleases
      tags: [tmdb]
      summary: Browse upcoming theatrical releases
      parameters:
        - name: region
          in: query
          schema:
            type: string
            default: "IN"
        - name: language
          in: query
          schema:
            type: string
            description: "Optional ISO 639-1 language filter (e.g. te, hi). Omit for Indian languages + English."
        - name: days
          in: query
          schema:
            type: integer
            default: 90
            minimum: 14
            maximum: 180
      responses:
        "200":
          description: Upcoming releases sorted by release date
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TmdbMovie"
  /tmdb/discover-indian:
    get:
      operationId: discoverIndian
      tags: [tmdb]
      summary: Discover Indian films by language
      parameters:
        - name: language
          in: query
          schema:
            type: string
            description: "Language code: te, ta, ml, kn, hi"
      responses:
        "200":
          description: Discovered movies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TmdbMovie"
  /tmdb/similar/{tmdbId}:
    get:
      operationId: getSimilarMovies
      tags: [tmdb]
      summary: Get movies similar to a given TMDB movie
      parameters:
        - name: tmdbId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Similar movies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TmdbMovie"
  /tmdb/watch-providers/{tmdbId}:
    get:
      operationId: getWatchProviders
      tags: [tmdb]
      summary: Get streaming/watch providers for a movie (India region)
      parameters:
        - name: tmdbId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Watch providers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WatchProviders"
  /tmdb/recommendations/{tmdbId}:
    get:
      operationId: getTmdbRecommendations
      tags: [tmdb]
      summary: Get TMDB recommendations for a movie
      parameters:
        - name: tmdbId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Recommended movies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TmdbMovie"
...
components:
  schemas:
...
    TmdbMovie:
      type: object
      required: [tmdbId, title]
      properties:
        tmdbId:
          type: integer
        title:
          type: string
        originalTitle:
          type: ["string", "null"]
        posterPath:
          type: ["string", "null"]
        releaseYear:
          type: ["integer", "null"]
        releaseDate:
          type: ["string", "null"]
          description: Release day YYYY-MM-DD when known
        originalLanguage:
          type: ["string", "null"]
        genres:
          type: ["array", "null"]
          items:
            type: string
        overview:
          type: ["string", "null"]
        voteAverage:
          type: ["number", "null"]
...
    WatchProvider:
      type: object
      required: [name]
      properties:
        providerId:
          type: integer
        name:
          type: string
        logoPath:
          type: ["string", "null"]
    WatchProviders:
      type: object
      required: [tmdbId]
      properties:
        tmdbId:
          type: integer
        flatrate:
          type: ["array", "null"]
          items:
            $ref: "#/components/schemas/WatchProvider"
        rent:
          type: ["array", "null"]
          items:
            $ref: "#/components/schemas/WatchProvider"
        buy:
          type: ["array", "null"]
          items:
            $ref: "#/components/schemas/WatchProvider"
        link:
          type: ["string", "null"]
...