# lib/api-zod/src/generated/types/tmdbMovie.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Indian Cinema Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface TmdbMovie {
  tmdbId: number;
  title: string;
  /** @nullable */
  originalTitle?: string | null;
  /** @nullable */
  posterPath?: string | null;
  /** @nullable */
  releaseYear?: number | null;
  /**
     * Release day YYYY-MM-DD when known
     * @nullable
     */
  releaseDate?: string | null;
  /** @nullable */
  originalLanguage?: string | null;
  /** @nullable */
  genres?: string[] | null;
  /** @nullable */
  overview?: string | null;
  /** @nullable */
  voteAverage?: number | null;
}